% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_pleio.R
\name{MR_pleio}
\alias{MR_pleio}
\title{\emph{\strong{in <5 words, what is this function's purpose?}}}
\usage{
MR_pleio(
  BetaOutcome,
  BetaExposure,
  SdOutcome,
  SdExposure,
  data,
  SignifThreshold,
  rho,
  methods = "IVW"
)
}
\arguments{
\item{BetaOutcome}{\emph{\strong{write what this is}}}

\item{BetaExposure}{\emph{\strong{write what this is}}}

\item{SdOutcome}{\emph{\strong{write what this is}}}

\item{SdExposure}{\emph{\strong{write what this is}}}

\item{data}{\emph{\strong{write what this is}}}

\item{SignifThreshold}{\emph{\strong{write what this is}}}

\item{rho}{\emph{\strong{write what this is}}}

\item{methods}{\emph{\strong{write what this is}}}
}
\description{
\emph{\strong{write a short description of what this function does}}
}
\examples{
MR_pleio()
}
\keyword{***what}
\keyword{do}
\keyword{keywords}
\keyword{want?***}
\keyword{you}
