% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MREgger.R
\name{MREgger}
\alias{MREgger}
\title{\emph{\strong{in <5 words, what is this function's purpose?}}}
\usage{
MREgger(BetaYG, BetaXG, seBetaYG, data)
}
\arguments{
\item{BetaYG}{\emph{\strong{write what this is}}}

\item{BetaXG}{\emph{\strong{write what this is}}}

\item{seBetaYG}{\emph{\strong{write what this is}}}

\item{data}{\emph{\strong{write what this is}}}
}
\description{
\emph{\strong{write a short description of what this function does}}
}
\examples{
MREgger()
}
\keyword{***what}
\keyword{do}
\keyword{keywords}
\keyword{want?***}
\keyword{you}
