% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchPleio.R
\name{SearchPleio}
\alias{SearchPleio}
\title{\emph{\strong{in <5 words, what is this function's purpose?}}}
\usage{
SearchPleio(
  BetaOutcome,
  BetaExposure,
  SdOutcome,
  SdExposure,
  data,
  rho,
  CausalEstimate,
  SdCausalEstimate
)
}
\arguments{
\item{BetaOutcome}{\emph{\strong{write what this is}}}

\item{BetaExposure}{\emph{\strong{write what this is}}}

\item{SdOutcome}{\emph{\strong{write what this is}}}

\item{SdExposure}{\emph{\strong{write what this is}}}

\item{data}{\emph{\strong{write what this is}}}

\item{rho}{\emph{\strong{write what this is}}}

\item{CausalEstimate}{\emph{\strong{write what this is}}}

\item{SdCausalEstimate}{\emph{\strong{write what this is}}}
}
\description{
\emph{\strong{write a short description of what this function does}}
}
\examples{
SearchPleio()
}
\keyword{***what}
\keyword{do}
\keyword{keywords}
\keyword{want?***}
\keyword{you}
